+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 #### next release (8.7.7)
 
+- Splitting the measurement tool into lines and polygons. It is now possible to add a point to an existing segment.
 - [The next improvement]
 
 #### 8.7.6 - 2024-08-22